wol-api/pyproject.toml
[project]
name = "wol-api"
version = "1.0.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.121.1",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.1",
    "wakeonlan>=3.1.0",
]
